Move Cosmetics out of Patches.py
